requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
supabase>=2.3.0
google-genai>=1.7.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
apscheduler>=3.10.4
httpx>=0.27.0
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
python-multipart>=0.0.24
